chatbot-app/backend/requirements.txt
# Core Framework
fastapi==0.116.1
uvicorn[standard]==0.35.0

# Strands Agent Framework
strands-agents==1.5.0
strands-agents-tools==0.2.3

# AWS & Bedrock
bedrock-agentcore==0.1.1
boto3>=1.40.1

# MCP & HTTP
mcp==1.12.3
requests==2.32.4
aiohttp==3.10.11

httpx>=0.24.0
anyio>=3.6.0

# System utilities
psutil>=5.8.0